requirements.txt
pandas
python-dotenv
Flask
beautifulsoup4
selenium
tqdm
fuzzywuzzy
python-Levenshtein
matplotlib
nltk
torch
transformers
google-generativeai
